backend/requirements.txt
astroid==2.0.4
autopep8==1.4.1
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-nested-routers==0.91
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
Pillow==5.3.0
psycopg2-binary==2.7.6
pycodestyle==2.4.0
PyJWT==1.7.1
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pytz==2018.5
six==1.11.0
wrapt==1.10.11
